pyproject.toml
[project]
name = "surprise-tree"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "biopython",
    "sentence-transformers",
    "pinecone",
    "tqdm",
    "python-dotenv",
]
